// StudioMystri/bms/frontend/types.ts
export interface Employee {
  id: string;
  name: string;
  roleId: string;
  email: string;
  phone: string;
  salary: number;
  joinDate: string;
  department: string;
  status: 'Active' | 'Leave' | 'Terminated' | 'Inactive';
  attendance: 'Present' | 'Absent' | 'Half-Day' | 'Late' | 'Leave';
  role?: string;
  leavePolicy?: number;
  leavesRemaining?: number;
  dob?: string;
  currentAddress?: string;
  permanentAddress?: string;
  bloodGroup?: string;
  emergencyContact?: string;
  idProof?: string;
  notes?: string;
  qualifications?: string;
  documents?: string[];
  credentials?: { username: string; initialPass: string };
}